make clippy happy, use u32 for target_block and not usize

+++ swap/src/cli/api.rs
@@ -521,7 +521,7 @@ async fn init_bitcoin_wallet(
             data_dir: data_dir.to_path_buf(),
         })
         .finality_confirmations(env_config.bitcoin_finality_confirmations)
-        .target_block(bitcoin_target_block as usize)
+        .target_block(bitcoin_target_block)
         .sync_interval(env_config.bitcoin_sync_interval())
         .env_config(env_config);
